# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Main API interfaces, including chat, session management, authentication, etc..",
        "title": "Morign API",
        "contact": {
            "name": "liut",
            "url": "https://github.com/liut"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/admin/skills": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "默认 文档生成"
                ],
                "summary": "查询 技能 列表 🔑",
                "operationId": "admin-skills-get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录票据凭证",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "aaa,bbb,ccc",
                        "x-order": "0",
                        "description": "主键编号`ids`（以逗号分隔的字串），仅供 Form 或 Query 使用, example:\"aaa,bbb,ccc\"",
                        "name": "ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "x-order": "2",
                        "description": "创建者ID",
                        "name": "creatorID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "x-order": "3",
                        "description": "创建时间 形式： yyyy-mm-dd, 1_day, 2_weeks, 3_months",
                        "name": "created",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "x-order": "4",
                        "description": "更新时间 形式： yyyy-mm-dd, 1_day, 2_weeks, 3_months",
                        "name": "updated",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "x-order": "5",
                        "description": "IsDelete 查询删除的记录",
                        "name": "isDelete",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "x-order": "A",
                        "description": "技能名 全局唯一（小写字母数字连字符）",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "x-order": "B",
                        "description": "可用频道 位掩码（0=未投放，仅创建者可见） (支持混合解码)\n * `web`\n * `wecom` - 企业微信\n * `feishu` - 飞书",
                        "name": "channel",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "x-order": "C",
                        "description": "创建者 uid",
                        "name": "owner",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "x-order": "D",
                        "description": "仅查看可见的",
                        "name": "visible",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "x-order": "[",
                        "description": "第几页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "x-order": "]",
                        "description": "跳过多少条记录，如果提供 page 此项跳过",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "x-order": "_",
                        "description": "分页大小，默认20",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "x-order": "|",
                        "description": "排序，允许最多两个字段排序 field1 [asc | desc] [,field2 [asc | desc] ...]",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Done"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/api.ResultData"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/skillsSkill"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求或参数错误",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    },
                    "404": {
                        "description": "目标未找到",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    },
                    "503": {
                        "description": "服务端错误",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "默认 文档生成"
                ],
                "summary": "录入 技能 🔑",
                "operationId": "admin-skills-post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录票据凭证",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Object",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/skillsSkillBasic"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Done"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/api.ResultID"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求或参数错误",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    },
                    "503": {
                        "description": "服务端错误",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    }
                }
            }
        },
        "/api/admin/skills/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "默认 文档生成"
                ],
                "summary": "获取 技能 详情 🔑",
                "operationId": "admin-skills-id-get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录票据凭证",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "编号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Done"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/skillsSkill"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求或参数错误",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    },
                    "404": {
                        "description": "目标未找到",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    },
                    "503": {
                        "description": "服务端错误",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "默认 文档生成"
                ],
                "summary": "更新 技能 🔑",
                "operationId": "admin-skills-id-put",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录票据凭证",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "编号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Object",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/skillsSkillSet"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Done"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求或参数错误",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    },
                    "503": {
                        "description": "服务端错误",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "默认 文档生成"
                ],
                "summary": "删除 技能 🔑",
                "operationId": "admin-skills-id-delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录票据凭证",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "编号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Done"
                        }
                    },
                    "400": {
                        "description": "请求或参数错误",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    },
                    "503": {
                        "description": "服务端错误",
                        "schema": {
                            "$ref": "#/definitions/api.Failure"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Meta": {
            "type": "object",
            "additionalProperties": {}
        },
...
        "api.Done": {
            "type": "object",
            "properties": {
                "extra": {
                    "description": "附加数据,可选"
                },
                "result": {
                    "description": "主体数据,可选"
                },
                "status": {
                    "description": "状态值，0=ok",
                    "type": "integer",
                    "example": 0
                },
                "t": {
                    "description": "时间戳",
                    "type": "integer"
                }
            }
        },
        "api.Failure": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "错误字段,可选,多用于表单校验",
                    "type": "string",
                    "example": "错误字段"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string",
                    "example": "错误信息"
                },
                "status": {
                    "description": "状态值",
                    "type": "integer",
                    "example": 1
                },
                "t": {
                    "description": "时间戳",
                    "type": "integer"
                }
            }
        },
...
        "api.ResultData": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "数据集数组"
                },
                "total": {
                    "description": "符合条件的总记录数",
                    "type": "integer"
                }
            }
        },
        "api.ResultID": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "主键值，多数时候是字串"
                }
            }
        },
...
        "skillsSkill": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "string",
                    "x-order": ""
                },
                "id": {
                    "description": "主键",
                    "type": "string",
                    "x-order": "!"
                },
                "updatedAt": {
                    "description": "变更时间",
                    "type": "string",
                    "x-order": "."
                },
                "name": {
                    "description": "技能名 全局唯一（小写字母数字连字符）",
                    "type": "string",
                    "x-order": "A"
                },
                "description": {
                    "description": "技能描述 做什么+何时用",
                    "type": "string",
                    "x-order": "B"
                },
                "content": {
                    "description": "正文",
                    "type": "string",
                    "x-order": "C"
                },
                "channel": {
                    "description": "可用频道 位掩码（0=未投放，仅创建者可见）\n * `web`\n * `wecom` - 企业微信\n * `feishu` - 飞书",
                    "type": "string",
                    "enum": [
                        "web",
                        "wecom",
                        "feishu"
                    ],
                    "x-order": "D"
                },
                "owner": {
                    "description": "创建者 uid",
                    "type": "string",
                    "x-order": "E"
                },
                "creatorID": {
                    "description": "创建者ID",
                    "type": "string",
                    "x-order": "_"
                },
                "meta": {
                    "description": "Meta 元信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Meta"
                        }
                    ],
                    "x-order": "|"
                }
            }
        },
        "skillsSkillBasic": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "description": "技能名 全局唯一（小写字母数字连字符）",
                    "type": "string",
                    "x-order": "A"
                },
                "description": {
                    "description": "技能描述 做什么+何时用",
                    "type": "string",
                    "x-order": "B"
                },
                "content": {
                    "description": "正文",
                    "type": "string",
                    "x-order": "C"
                },
                "channel": {
                    "description": "可用频道 位掩码（0=未投放，仅创建者可见）\n * `web`\n * `wecom` - 企业微信\n * `feishu` - 飞书",
                    "type": "string",
                    "enum": [
                        "web",
                        "wecom",
                        "feishu"
                    ],
                    "x-order": "D"
                },
                "owner": {
                    "description": "创建者 uid",
                    "type": "string",
                    "x-order": "E"
                }
            }
        },
        "skillsSkillSet": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "技能名 全局唯一（小写字母数字连字符）",
                    "type": "string",
                    "x-order": "A"
                },
                "description": {
                    "description": "技能描述 做什么+何时用",
                    "type": "string",
                    "x-order": "B"
                },
                "content": {
                    "description": "正文",
                    "type": "string",
                    "x-order": "C"
                },
                "channel": {
                    "description": "可用频道 位掩码（0=未投放，仅创建者可见）\n * `web`\n * `wecom` - 企业微信\n * `feishu` - 飞书",
                    "type": "string",
                    "enum": [
                        "web",
                        "wecom",
                        "feishu"
                    ],
                    "x-order": "D"
                }
            }
        }
    }
}